U.I Improvements
This push added a new background colour to each diagnostic, some improvement to the button as well as a tooltip for one of the questions in the lung cancer diagnoses, future push will be on the tooltip.
minimum window size set, although full-screen window to be done, application icon implemented, the .icon is from https://en.m.wikipedia.org/wiki/File:Circle-icons-clipboard.svg.
future U.I improvements to focus on functionality as well to sucess the user what to do with tell him directly only through design q.
Also User testing as done for this update, done by me with the help of a friend named ****** *********, making it possible for more detail and comprehensive user testing analysis, The user consent to reveal his name, but still idk if I should revel it in a GitHub push